+++ tools/test_sources.gd
@@ -23,8 +23,9 @@ func _init() -> void:
 	print("ok=", r3.ok, " ", (r3.texture.get_size() if r3.ok else r3.error))
 
 	print("--- 4. external engraver (inline MusicXML via stub) ---")
+	var stub := ProjectSettings.globalize_path("res://tools/stub_engraver.py")
 	ProjectSettings.set_setting("musicscene/notation/engraver/musicxml",
-		'py "D:/Projects/MusicScene/tools/stub_engraver.py" {input} {output} {format}')
+		'py "%s" {input} {output} {format}' % stub)
 	var mxml := '<?xml version="1.0"?><score-partwise><part-list/></score-partwise>'
 	var r4 = R.render(mxml, "musicxml", 1)
 	print("ok=", r4.ok, " ", (r4.texture.get_size() if r4.ok else r4.error))